feat: collection endpoints create

+++ client/src/actions/items.js
@@ -1,11 +1,20 @@
-import { FETCH_ALL, FETCH_BY_SEARCH, FETCH_ITEM, CREATE, UPDATE, DELETE, LIKE, START_LOADING, END_LOADING, COMMENT} from '../components/constants/actionTypes';
+import { FETCH_ALL, FETCH_BY_SEARCH, FETCH_ITEM, CREATE, UPDATE, DELETE, LIKE, START_LOADING, END_LOADING, COMMENT, BASE_URL} from '../components/constants/actionTypes';
 import * as api from '../api';
 
+function setHeaders(headers) {
+    if(localStorage.getItem('profile')) {
+        return {
+            ...headers,
+            'Authorization': `Bearer ${JSON.parse(localStorage.getItem('profile')).token}`
+        }
+    } else {
+        return headers;
+    }
+}
 export const getItem = (id) =>  async (dispatch) => {
     try {
         dispatch({type: START_LOADING});
         const { data } = await api.fetchItem(id);
-        console.log(data);
         dispatch({ type: FETCH_ITEM, payload: data });
         dispatch({type: END_LOADING});
     } catch (error) {
@@ -13,11 +22,15 @@ export const getItem = (id) =>  async (dispatch) => {
     }
 }
 
-export const getItems = (page) =>  async (dispatch) => {
+export const getItems = () =>  async (dispatch) => {
     try {
         dispatch({type: START_LOADING});
-        const { data } = await api.fetchItems(page);
-        dispatch({ type: FETCH_ALL, payload: data });
+        const req = await fetch("http://localhost:5000/items",{
+            headers: setHeaders({})});
+        const data = await req.json();
+        if(data.status === "ok"){ 
+            dispatch({ type: FETCH_ALL, payload: data });
+        }
         dispatch({type: END_LOADING});
     } catch (error) {
         console.log(error); 
@@ -38,13 +51,22 @@ export const getItemsBySearch = (searchQuery) => async (dispatch) => {
 }
 
 export const createItem = (item) => async (dispatch) => {
-    try{
+    try {
         dispatch({type: START_LOADING});
-        const { data } = await api.createItem(item);
-        dispatch({ type: CREATE, payload: data });
+        const req = await fetch("http://localhost:5000/items",{
+            headers: setHeaders({
+                'Content-Type': 'application/json',
+            }),
+            method: 'POST',
+            body: JSON.stringify(item)
+        });
+        const data = await req.json()
+        if(data.status === "ok"){ 
+            dispatch({ type: CREATE, payload: data.newItem });
+        }
         dispatch({type: END_LOADING});
     } catch (error) {
-        console.log(error)
+        console.log(error); 
     }
 }
 
@@ -60,8 +82,16 @@ export const updateItem = (id, item) => async (dispatch) => {
 
 export const deleteItem = (id) => async (dispatch) => {
     try{
-        await api.deleteItem(id);
-        dispatch({ type: DELETE, payload: id });
+        const response = await fetch(`http://localhost:5000/items/${id}`,{
+            method: 'DELETE', 
+            headers: setHeaders({
+                'Content-Type': 'application/json',
+            }),
+        })
+        const data = await response.json();
+        if(data.status === 'ok' ){ 
+            dispatch({ type: DELETE, payload: id });
+        }
     } catch (error) {
         console.log(error)
     }
@@ -69,10 +99,21 @@ export const deleteItem = (id) => async (dispatch) => {
 
 export const likeItem = (id) => async (dispatch) => {
     try{
-        const { data } = await api.likeItem(id);
-
-        dispatch({ type: LIKE, payload: data });
+        const response = await fetch(`http://localhost:5000/items/${id}/likeItem`,{
+            method: 'PUT', 
+            headers: setHeaders({
+                'Content-Type': 'application/json',
+            }),
+        })
+        console.log(response);
+        const data = await response.json();
+        console.log(data);
+        if(data.status === 'ok' ) {
+            dispatch({ type: LIKE, payload: data.updatedItem });
+        }
     } catch (error) {
+        console.log('HERE IT IS');
+        
         console.log(error)
     }
 }
